Renamed the enum.
